Restructured api tests
